build(image): support multiarch container images build for docker and other build tools. (#2)
This pull request introduces significant changes to the build workflows,
containerization process, and project structure to support
multi-architecture container images and enhance the Gradle build logic.
The most important updates include the addition of multi-architecture
support in build workflows, the creation of a custom Gradle plugin for
container tasks, and the replacement of the `Dockerfile` with a
`Containerfile`.

### Highlights:
* Added support for multi-architecture image builds
(`buildImageMultiArch`) and packaging (`packageImageMultiArch`). Updated
the image push process to include architecture-specific tags and a
manifest for multi-arch images.
* Introduced a new `Containerfile` to replace the `Dockerfile`.
* Added a custom Gradle plugin
(`tech.lamprism.lampray.gradle.container.ContainerPlugin`) for
container-related tasks. Add a task for building OCI-compliant container
images using customizable tools and build contexts.